Contractors module refactored
