+++ apps/frontend/src/app/app.component.html
@@ -65,7 +65,7 @@
         <mat-icon matListItemIcon> article </mat-icon>
         <div matListItemTitle>Проформа фактури</div>
       </mat-list-item>
-      <mat-list-item matRipple routerLink="/contractorsList">
+      <mat-list-item matRipple routerLink="/contractors">
         <mat-icon matListItemIcon> handshake </mat-icon>
         <div matListItemTitle>Контрагенти</div>
       </mat-list-item>
