+++ apps/frontend/src/app/modules/contractors/contractor/contractor.component.scss
@@ -37,7 +37,6 @@ mat-form-field {
   mat-card-actions {
     justify-content: center;
   }
-
 }
 
 @media screen and (max-width: 300px) {
