+++ apps/frontend/src/app/modules/contractors/contractor/contractor.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ContractorComponent } from './contractor.component';
+
+describe('ContractorComponent', () => {
+  let component: ContractorComponent;
+  let fixture: ComponentFixture<ContractorComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ContractorComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(ContractorComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
