+++ apps/frontend/src/app/modules/contractors/contractors.component.html
@@ -1,108 +1,66 @@
-<form [formGroup]="contractorsForm" (ngSubmit)="onSubmit()">
-  <mat-card class="contactor-card">
-    <mat-card-header>
-      <mat-card-title>Контрагент</mat-card-title>
-    </mat-card-header>
-    <br />
-    <mat-card-content>
-      <p>
-        <mat-form-field appearance="outline">
-          <mat-label>Име</mat-label>
-          <input matInput formControlName="name" />
+<div class="mat-elevation-z8">
+  <h1>Контрагенти</h1>
+  <button
+    class="btn-add"
+    mat-raised-button
+    matTooltip="добавяне на контрагент"
+    aria-label="добавяне на контрагент"
+  >
+    <a routerLink="./new">Добави</a>
+  </button>
 
-          <mat-hint [ngStyle]="{ color: '#f44336' }" *ngIf="isFormInvalid()"
-            >Полетата отбелязани с * са задължителни!</mat-hint
-          >
-          <mat-error *ngIf="contractorsForm.get('name')?.errors?.['minlength']"
-            >Името трябва да е поне 2 знака!</mat-error
-          >
-        </mat-form-field>
-      </p>
-      <p>
-        <mat-checkbox
-          color="primary"
-          class="example-margin"
-          formControlName="person"
-          >Физическо лице</mat-checkbox
+  <table mat-table [dataSource]="dataSource" class="">
+    <h2>Контрагенти</h2>
+    <!-- Name Column -->
+    <ng-container matColumnDef="name">
+      <th mat-header-cell *matHeaderCellDef>Име</th>
+      <td mat-cell *matCellDef="let contractor">{{ contractor.name }}</td>
+    </ng-container>
+
+    <!-- City Column -->
+    <ng-container matColumnDef="city">
+      <th mat-header-cell *matHeaderCellDef>Град</th>
+      <td mat-cell *matCellDef="let contractor">{{ contractor.city }}</td>
+    </ng-container>
+
+    <!-- MOL Column -->
+    <ng-container matColumnDef="mol">
+      <th mat-header-cell *matHeaderCellDef>МОЛ</th>
+      <td mat-cell *matCellDef="let contractor">{{ contractor.mol }}</td>
+    </ng-container>
+
+    <!-- EGN Column -->
+    <ng-container matColumnDef="egn">
+      <th mat-header-cell *matHeaderCellDef>ЕИК/ЕГН</th>
+      <td mat-cell *matCellDef="let contractor">
+        {{ contractor.egn ? contractor.egn : contractor.eik }}
+      </td>
+    </ng-container>
+
+    <!-- Tools Column -->
+    <ng-container matColumnDef="tools">
+      <th mat-header-cell *matHeaderCellDef>Инструменти</th>
+      <td mat-cell *matCellDef="let contractor">
+        <button
+          class="icon-edit-btn"
+          mat-icon-button
+          (click)="editContractor(contractor)"
+          matTooltip="редактиране на ред"
+          aria-label="редактиране на ред"
         >
-      </p>
+          <mat-icon>edit</mat-icon>
+        </button>
+      </td>
+    </ng-container>
 
-      <p id="contractorEik">
-        <mat-form-field appearance="outline">
-          <mat-label>ЕИК</mat-label>
-          <input matInput formControlName="eik" />
-          <mat-error *ngIf="contractorsForm.get('eik')?.errors?.['invalidEik']"
-            >ЕИК номерът трябва да се състои от 9 или 13 цифри.</mat-error
-          >
-        </mat-form-field>
-      </p>
-      <p id="contractorEgn" class="hidden">
-        <mat-form-field appearance="outline">
-          <mat-label>ЕГН</mat-label>
-          <input matInput formControlName="egn" />
-          <mat-error
-            *ngIf="contractorsForm.get('egn')?.errors?.['invalidEgnLength']"
-            >ЕГН Трябва да се състои от 10 знака!</mat-error
-          >
-          <mat-error
-            *ngIf="contractorsForm.get('egn')?.errors?.['invalidEgnChecksum']"
-            >Невалидно ЕГН!</mat-error
-          >
-        </mat-form-field>
-      </p>
-      <p id="contractorVatNumber">
-        <mat-form-field appearance="outline">
-          <mat-label>ДДС Номер</mat-label>
-          <input matInput formControlName="ddsnumber" />
-        </mat-form-field>
-      </p>
-      <p id="contractorMol">
-        <mat-form-field appearance="outline">
-          <mat-label>МОЛ</mat-label>
-          <input matInput formControlName="mol" />
-
-          <mat-error *ngIf="contractorsForm.get('mol')?.hasError('minlength')"
-            >МОЛ трябва да съдържа поне 2 букви!</mat-error
-          >
-          <mat-error *ngIf="contractorsForm.get('mol')?.hasError('maxlength')"
-            >МОЛ трябва да съдържа максимум 40 букви!</mat-error
-          >
-        </mat-form-field>
-      </p>
-      <p>
-        <mat-form-field appearance="outline">
-          <mat-label>Населено място</mat-label>
-          <input matInput formControlName="city" />
-          <ng-template #minLettersError>
-            <mat-error
-              *ngIf="contractorsForm.get('city')?.hasError('minlength')"
-              >Населено място трябва да съдържа минимум 3 букви!</mat-error
-            >
-          </ng-template>
-          <mat-error
-            *ngIf="
-              contractorsForm.get('city')?.hasError('pattern');
-              else minLettersError
-            "
-            >Населено място трябва да започва с главна буква!</mat-error
-          >
-        </mat-form-field>
-      </p>
-      <p>
-        <mat-form-field appearance="outline">
-          <mat-label>Адрес</mat-label>
-          <input matInput formControlName="address" />
-          <mat-error
-            *ngIf="contractorsForm.get('address')?.hasError('minlength')"
-            >Адресът трябва да съдържа поне 5 символа!</mat-error
-          >
-        </mat-form-field>
-      </p>
-    </mat-card-content>
-    <mat-card-actions>
-      <button [disabled]="isFormInvalid()" type="submit" mat-raised-button>
-        <a>Изпрати</a>
-      </button>
-    </mat-card-actions>
-  </mat-card>
-</form>
+    <!-- Table rows -->
+    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+    <tr mat-row *matRowDef="let contractor; columns: displayedColumns"></tr>
+  </table>
+  <mat-paginator
+    [length]="contractors.length"
+    [pageSize]="5"
+    [pageSizeOptions]="[5, 10, 25, 100]"
+    showFirstLastButtons
+  ></mat-paginator>
+</div>
