+++ apps/frontend/src/app/modules/contractors/contractors.component.scss
