+++ apps/frontend/src/app/modules/contractors/contractors.module.ts
@@ -6,18 +6,21 @@ import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
-import { MatDatepickerModule } from '@angular/material/datepicker';
-import { MatSliderModule } from '@angular/material/slider';
-import { MatMenuModule } from '@angular/material/menu';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatNativeDateModule } from '@angular/material/core';
 import { ReactiveFormsModule } from '@angular/forms';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatTableModule } from '@angular/material/table';
+import { ContractorComponent } from './contractor/contractor.component';
 
-const routes: Routes = [{ path: '', component: ContractorsComponent }];
+const routes: Routes = [
+  { path: '', component: ContractorsComponent },
+  { path: 'new', component: ContractorComponent },
+  { path: ':id', component: ContractorComponent },
+];
 
 @NgModule({
-  declarations: [ContractorsComponent],
+  declarations: [ContractorsComponent, ContractorComponent],
   imports: [
     CommonModule,
     RouterModule.forChild(routes),
@@ -25,13 +28,11 @@ const routes: Routes = [{ path: '', component: ContractorsComponent }];
     MatButtonModule,
     MatFormFieldModule,
     MatInputModule,
-    MatDatepickerModule,
-    MatSliderModule,
-    MatMenuModule,
     MatIconModule,
     MatCheckboxModule,
-    MatNativeDateModule,
     ReactiveFormsModule,
+    MatPaginatorModule,
+    MatTableModule,
   ],
 })
 export class ContractorsModule {}
